program as it is before a major change to use vectors instead of angles.
I am doing this because a lot of the maths become significantly simpler
with vectors. Also, steering behaviors use vectors to represent how much
force is being applied to the boid during movement calculations. This
will help improve the program as well since vectors have a magnitude and
a direction, where as angles only have a direction.